move vpt2 into procedures/
